refactor (XSLT): make utility function for XSLT
